+++ example/lib/main.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 
 import 'package:device_policy_controller/device_policy_controller.dart';
 
+import 'task_actions.dart';
+
 void main() {
   runApp(const MyApp());
 }
@@ -22,154 +24,57 @@ class HomePage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final screenAwakeValuesNotifier = ValueNotifier([true, false]);
+    final screenAwakeValueNotifier = ValueNotifier(false);
+
+    final screenSizeWidth = MediaQuery.sizeOf(context).width;
 
     return Scaffold(
       appBar: AppBar(
         title: const Text('Device Policy Controller Example'),
+        actions: [
+          ValueListenableBuilder(
+            valueListenable: screenAwakeValueNotifier,
+            builder: (context, val, _) {
+              return TextButton(
+                onPressed: () {
+                  DevicePolicyController.instance
+                      .setKeepScreenAwake(!screenAwakeValueNotifier.value)
+                      .then(
+                    (value) {
+                      screenAwakeValueNotifier.value =
+                          !screenAwakeValueNotifier.value;
+                    },
+                  );
+                },
+                child: Icon(Icons.flash_on,
+                    color: val ? Colors.amberAccent : Colors.blueGrey,
+                    size: 36),
+              );
+            },
+          ),
+          const SizedBox(width: 12),
+        ],
       ),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            ElevatedButton(
-              onPressed: () {
-                DevicePolicyController.instance
-                    .requestAdminPrivilegesIfNeeded()
-                    .then(
-                  (isGranted) {
-                    showDialog(
-                      context: context,
-                      builder: (context) {
-                        return AlertDialog(
-                          title: const Text("Requests admin privileges"),
-                          content: isGranted
-                              ? const Text("Admin privileges is granted.")
-                              : const Text(
-                                  "Admin privileges have not been granted.\n"
-                                  "Either the user has declined the request or This app is not a Device Policy Controller (DPC)."),
-                          actions: [
-                            TextButton(
-                              onPressed: () {
-                                Navigator.pop(context);
-                              },
-                              child: const Text("OK"),
-                            ),
-                          ],
-                        );
-                      },
-                    );
-                  },
-                );
-              },
-              child: const Text('Requests admin privileges if needed.'),
-            ),
-            ElevatedButton(
-              onPressed: () {
-                DevicePolicyController.instance.lockApp();
-              },
-              child: const Text('Locks the app in kiosk mode'),
-            ),
-            ElevatedButton(
-              onPressed: () {
-                DevicePolicyController.instance.unlockApp();
-              },
-              child: const Text('Unlocks the app.'),
-            ),
-            ElevatedButton(
-              onPressed: () {
-                DevicePolicyController.instance.isAdminActive().then((isAdmin) {
-                  showDialog(
-                    context: context,
-                    builder: (context) {
-                      return AlertDialog(
-                        title: const Text("Admin Privileges"),
-                        content: isAdmin
-                            ? const Text("The app has admin privileges.")
-                            : const Text(
-                                "The app does not have admin privileges."),
-                        actions: [
-                          TextButton(
-                            onPressed: () {
-                              Navigator.pop(context);
-                            },
-                            child: const Text("OK"),
-                          ),
-                        ],
-                      );
-                    },
-                  );
-                });
-              },
-              child: const Text('Checks if admin privileges are active.'),
-            ),
-            ElevatedButton(
-              onPressed: () {
-                DevicePolicyController.instance
-                    .getDeviceInfo()
-                    .then((Map<String, dynamic>? info) {
-                  showDialog(
-                    context: context,
-                    builder: (context) {
-                      return AlertDialog(
-                        title: const Text("Device Information"),
-                        content: info != null
-                            ? Column(
-                                crossAxisAlignment: CrossAxisAlignment.start,
-                                mainAxisSize: MainAxisSize.min,
-                                children: info.entries
-                                    .map((i) => Text("${i.key}: ${i.value}"))
-                                    .toList(),
-                              )
-                            : const Text(
-                                "Unable to retrieve device information."),
-                        actions: [
-                          TextButton(
-                            onPressed: () {
-                              Navigator.pop(context);
-                            },
-                            child: const Text("OK"),
-                          ),
-                        ],
-                      );
-                    },
-                  );
-                });
-              },
-              child: const Text('Gets device information'),
-            ),
-            const SizedBox(height: 32),
-            const Text("Keep Screen Awake ?"),
-            const SizedBox(height: 8),
-            ValueListenableBuilder(
-              valueListenable: screenAwakeValuesNotifier,
-              builder: (context, val, _) {
-                return ToggleButtons(
-                  onPressed: (index) {
-                    DevicePolicyController.instance
-                        .setKeepScreenAwake(index == 1)
-                        .then(
-                      (value) {
-                        screenAwakeValuesNotifier.value = [
-                          index == 0,
-                          index == 1
-                        ];
-                      },
-                    );
-                  },
-                  isSelected: val,
-                  children: const [
-                    Padding(
-                      padding: EdgeInsets.all(8.0),
-                      child: Icon(Icons.flash_off, size: 56),
-                    ),
-                    Padding(
-                      padding: EdgeInsets.all(8.0),
-                      child: Icon(Icons.flash_on, size: 56),
-                    )
-                  ],
-                );
-              },
+            Expanded(
+              child: GridView.builder(
+                padding: const EdgeInsets.all(16.0),
+                gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+                  crossAxisCount: 2,
+                  crossAxisSpacing: 16.0,
+                  mainAxisSpacing: 16.0,
+                  childAspectRatio:
+                      (screenSizeWidth / 2) / (screenSizeWidth / 8),
+                ),
+                itemCount: taskActions.length,
+                itemBuilder: (context, index) {
+                  final action = taskActions[index];
+                  return action.button(context);
+                },
+              ),
             ),
           ],
         ),