+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.2
+
+Add methods:
+* clearDeviceOwnerApp({String? packageName})
+* setCameraDisabled({required bool disabled})
+* setKeyguardDisabled({required bool disabled})
+* setScreenCaptureDisabled({required bool disabled})
+
+
 ## 0.0.1
 
 * Initial version.